+++ tech1-framework-domain/src/main/java/io/tech1/framework/domain/properties/configs/IncidentConfigs.java
@@ -3,23 +3,21 @@ package io.tech1.framework.domain.properties.configs;
 import io.tech1.framework.domain.properties.annotations.MandatoryProperty;
 import io.tech1.framework.domain.properties.annotations.NonMandatoryProperty;
 import io.tech1.framework.domain.properties.base.RemoteServer;
-import io.tech1.framework.domain.properties.configs.incidents.AbstractIncidentFeatureConfigs;
 import io.tech1.framework.domain.properties.configs.incidents.IncidentFeaturesConfigs;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
 import static java.lang.Boolean.TRUE;
-import static java.util.Objects.isNull;
 
 // Lombok (property-based)
 @Data
 @EqualsAndHashCode(callSuper = true)
-public class IncidentConfigs extends AbstractPropertiesToggleConfigs implements AbstractIncidentFeatureConfigs {
+public class IncidentConfigs extends AbstractPropertiesToggleConfigs {
     @MandatoryProperty
     private boolean enabled;
-    @NonMandatoryProperty
+    @MandatoryProperty
     private RemoteServer remoteServer;
-    @NonMandatoryProperty
+    @MandatoryProperty
     private IncidentFeaturesConfigs features;
 
     // NOTE: test-purposes
@@ -52,13 +50,5 @@ public class IncidentConfigs extends AbstractPropertiesToggleConfigs implements 
         if (TRUE.equals(loginFailureUsernamePassword.isEnabled()) && TRUE.equals(loginFailureUsernameMaskedPassword.isEnabled())) {
             throw new IllegalArgumentException("Please configure login failure incident feature. Only one feature type could be enabled");
         }
-    }
-
-    @Override
-    public void configureRequiredIncidentsIfMissing() {
-        if (isNull(this.features)) {
-            this.features = new IncidentFeaturesConfigs();
-        }
-        this.features.configureRequiredIncidentsIfMissing();
     }
 }