+++ tech1-framework-domain/src/main/java/io/tech1/framework/domain/properties/configs/incidents/AbstractIncidentFeatureConfigs.java
@@ -1,5 +1,0 @@
-package io.tech1.framework.domain.properties.configs.incidents;
-
-public interface AbstractIncidentFeatureConfigs {
-    void configureRequiredIncidentsIfMissing();
-}