+++ tech1-framework-domain/src/test/java/io/tech1/framework/domain/properties/configs/IncidentConfigsTest.java
@@ -5,9 +5,6 @@ import io.tech1.framework.domain.properties.configs.incidents.IncidentFeaturesCo
 import io.tech1.framework.domain.utilities.random.EntityUtility;
 import org.junit.jupiter.api.Test;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import static io.tech1.framework.domain.properties.configs.incidents.IncidentFeatureConfigs.disabledIncidentFeatureConfigs;
 import static io.tech1.framework.domain.properties.configs.incidents.IncidentFeatureConfigs.enabledIncidentFeatureConfigs;
 import static io.tech1.framework.domain.utilities.reflections.ReflectionUtility.getGetters;
@@ -77,54 +74,6 @@ public class IncidentConfigsTest {
     }
 
     @Test
-    public void configureRequiredIncidentsTest() {
-        List<IncidentFeaturesConfigs> cases = new ArrayList<>();
-        cases.add(new IncidentFeaturesConfigs());
-        cases.add(null);
-        cases.forEach(incidentFeaturesConfigsParam -> {
-            // Arrange
-            var incidentConfigs = IncidentConfigs.of(
-                    true,
-                    EntityUtility.entity(RemoteServer.class),
-                    incidentFeaturesConfigsParam
-            );
-
-            // Act
-            incidentConfigs.configureRequiredIncidentsIfMissing();
-
-            // Assert
-            assertThat(incidentConfigs.isEnabled()).isTrue();
-            assertThat(incidentConfigs.getRemoteServer()).isNotNull();
-            var incidentFeaturesConfigs = incidentConfigs.getFeatures();
-            assertThat(incidentFeaturesConfigs).isNotNull();
-            assertThat(getGetters(incidentFeaturesConfigs)).hasSize(8);
-            assertThat(incidentFeaturesConfigs.getLogin()).isNotNull();
-            assertThat(incidentFeaturesConfigs.getLogin().isEnabled()).isFalse();
-
-            assertThat(incidentFeaturesConfigs.getLoginFailureUsernamePassword()).isNotNull();
-            assertThat(incidentFeaturesConfigs.getLoginFailureUsernamePassword().isEnabled()).isFalse();
-
-            assertThat(incidentFeaturesConfigs.getLoginFailureUsernameMaskedPassword()).isNotNull();
-            assertThat(incidentFeaturesConfigs.getLoginFailureUsernameMaskedPassword().isEnabled()).isFalse();
-
-            assertThat(incidentFeaturesConfigs.getLogout()).isNotNull();
-            assertThat(incidentFeaturesConfigs.getLogout().isEnabled()).isFalse();
-
-            assertThat(incidentFeaturesConfigs.getSessionRefreshed()).isNotNull();
-            assertThat(incidentFeaturesConfigs.getSessionRefreshed().isEnabled()).isFalse();
-
-            assertThat(incidentFeaturesConfigs.getSessionExpired()).isNotNull();
-            assertThat(incidentFeaturesConfigs.getSessionExpired().isEnabled()).isFalse();
-
-            assertThat(incidentFeaturesConfigs.getRegister1()).isNotNull();
-            assertThat(incidentFeaturesConfigs.getRegister1().isEnabled()).isFalse();
-
-            assertThat(incidentFeaturesConfigs.getRegister1Failure()).isNotNull();
-            assertThat(incidentFeaturesConfigs.getRegister1Failure().isEnabled()).isFalse();
-        });
-    }
-
-    @Test
     public void notConfigureRequiredIncidentsTest() {
         // Arrange
         var incidentConfigs = IncidentConfigs.of(
@@ -138,13 +87,13 @@ public class IncidentConfigsTest {
                         enabledIncidentFeatureConfigs(),
                         enabledIncidentFeatureConfigs(),
                         enabledIncidentFeatureConfigs(),
+                        enabledIncidentFeatureConfigs(),
                         enabledIncidentFeatureConfigs()
                 )
         );
 
         // Act
         incidentConfigs.assertProperties();
-        incidentConfigs.configureRequiredIncidentsIfMissing();
 
         // Assert
         assertThat(incidentConfigs.isEnabled()).isTrue();
@@ -163,6 +112,9 @@ public class IncidentConfigsTest {
 
         assertThat(incidentFeaturesConfigs.getLogout()).isNotNull();
         assertThat(incidentFeaturesConfigs.getLogout().isEnabled()).isTrue();
+
+        assertThat(incidentFeaturesConfigs.getLogoutMin()).isNotNull();
+        assertThat(incidentFeaturesConfigs.getLogoutMin().isEnabled()).isTrue();
 
         assertThat(incidentFeaturesConfigs.getSessionRefreshed()).isNotNull();
         assertThat(incidentFeaturesConfigs.getSessionRefreshed().isEnabled()).isTrue();