+++ tech1-framework-incidents/src/main/java/io/tech1/framework/incidents/events/publishers/impl/IncidentPublisherImpl.java
@@ -2,7 +2,6 @@ package io.tech1.framework.incidents.events.publishers.impl;
 
 import io.tech1.framework.domain.pubsub.AbstractEventPublisher;
 import io.tech1.framework.incidents.domain.Incident;
-import io.tech1.framework.incidents.domain.IncidentAttributes;
 import io.tech1.framework.incidents.domain.authetication.*;
 import io.tech1.framework.incidents.domain.registration.IncidentRegistration1;
 import io.tech1.framework.incidents.domain.registration.IncidentRegistration1Failure;
@@ -18,6 +17,7 @@ import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 
 import static io.tech1.framework.domain.constants.FrameworkLogsConstants.*;
+import static io.tech1.framework.incidents.domain.IncidentAttributes.IncidentsTypes.*;
 
 @Slf4j
 @Service
@@ -35,7 +35,7 @@ public class IncidentPublisherImpl extends AbstractEventPublisher implements Inc
             LOGGER.debug(INCIDENT_AUTHENTICATION_LOGIN, this.getType(), incident.getUsername());
             this.applicationEventPublisher.publishEvent(incident);
         } else {
-            LOGGER.warn(INCIDENT_FEATURE_DISABLED, IncidentAttributes.IncidentsTypes.AUTHENTICATION_LOGIN);
+            LOGGER.warn(INCIDENT_FEATURE_DISABLED, AUTHENTICATION_LOGIN);
         }
     }
 
@@ -57,11 +57,11 @@ public class IncidentPublisherImpl extends AbstractEventPublisher implements Inc
 
     @Override
     public void publishAuthenticationLogoutMin(IncidentAuthenticationLogoutMin incident) {
-        if (this.applicationFrameworkProperties.getIncidentConfigs().getFeatures().getLogout().isEnabled()) {
+        if (this.applicationFrameworkProperties.getIncidentConfigs().getFeatures().getLogoutMin().isEnabled()) {
             LOGGER.debug(INCIDENT_AUTHENTICATION_LOGOUT, this.getType(), incident.getUsername());
             this.applicationEventPublisher.publishEvent(incident);
         } else {
-            LOGGER.warn(INCIDENT_FEATURE_DISABLED, IncidentAttributes.IncidentsTypes.AUTHENTICATION_LOGOUT);
+            LOGGER.warn(INCIDENT_FEATURE_DISABLED, AUTHENTICATION_LOGOUT_MIN);
         }
     }
 
@@ -71,7 +71,7 @@ public class IncidentPublisherImpl extends AbstractEventPublisher implements Inc
             LOGGER.debug(INCIDENT_AUTHENTICATION_LOGOUT, this.getType(), incident.getUsername());
             this.applicationEventPublisher.publishEvent(incident);
         } else {
-            LOGGER.warn(INCIDENT_FEATURE_DISABLED, IncidentAttributes.IncidentsTypes.AUTHENTICATION_LOGOUT);
+            LOGGER.warn(INCIDENT_FEATURE_DISABLED, AUTHENTICATION_LOGOUT);
         }
     }
 
@@ -81,7 +81,7 @@ public class IncidentPublisherImpl extends AbstractEventPublisher implements Inc
             LOGGER.debug(INCIDENT_REGISTER1, this.getType(), incident.getUsername());
             this.applicationEventPublisher.publishEvent(incident);
         } else {
-            LOGGER.warn(INCIDENT_FEATURE_DISABLED, IncidentAttributes.IncidentsTypes.REGISTER1);
+            LOGGER.warn(INCIDENT_FEATURE_DISABLED, REGISTER1);
         }
     }
 
@@ -91,7 +91,7 @@ public class IncidentPublisherImpl extends AbstractEventPublisher implements Inc
             LOGGER.debug(INCIDENT_REGISTER1_FAILURE, this.getType(), incident.getUsername());
             this.applicationEventPublisher.publishEvent(incident);
         } else {
-            LOGGER.warn(INCIDENT_FEATURE_DISABLED, IncidentAttributes.IncidentsTypes.REGISTER1_FAILURE);
+            LOGGER.warn(INCIDENT_FEATURE_DISABLED, REGISTER1_FAILURE);
         }
     }
 
@@ -101,7 +101,7 @@ public class IncidentPublisherImpl extends AbstractEventPublisher implements Inc
             LOGGER.debug(INCIDENT_SESSION_REFRESHED, this.getType(), incident.getUsername());
             this.applicationEventPublisher.publishEvent(incident);
         } else {
-            LOGGER.warn(INCIDENT_FEATURE_DISABLED, IncidentAttributes.IncidentsTypes.SESSION_REFRESHED);
+            LOGGER.warn(INCIDENT_FEATURE_DISABLED, SESSION_REFRESHED);
         }
     }
 
@@ -111,7 +111,7 @@ public class IncidentPublisherImpl extends AbstractEventPublisher implements Inc
             LOGGER.debug(INCIDENT_SESSION_EXPIRED, this.getType(), incident.getUsername());
             this.applicationEventPublisher.publishEvent(incident);
         } else {
-            LOGGER.warn(INCIDENT_FEATURE_DISABLED, IncidentAttributes.IncidentsTypes.SESSION_EXPIRED);
+            LOGGER.warn(INCIDENT_FEATURE_DISABLED, SESSION_EXPIRED);
         }
     }